+++ src/modules/index.ts
@@ -9,8 +9,7 @@ import "./burgs-generator";
 import "./biomes";
 import "./cultures-generator";
 import "./routes-generator";
-import "./journey-draw";
-import "./journey-editor";
+import "./journey";
 import "./states-generator";
 import "./zones-generator";
 import "./religions-generator";
